fix: render email as plaintext
